% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedonadbplyr-package.R
\docType{package}
\name{sedonadbplyr-package}
\alias{sedonadbplyr}
\alias{sedonadbplyr-package}
\title{sedonadbplyr: Apache SedonaDB 'dplyr' Interface}
\description{
Provides a 'dplyr' backend implemented using 'SedonaDB'.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca}

}
\keyword{internal}
